+++ tests/test_get_version.py
@@ -4,12 +4,12 @@ import os
 import unittest
 from unittest import mock
 
-from pyobo.api.utils import get_version, get_version_pins
 from pyobo.utils.misc import (
     _get_version_from_artifact,
     _prioritize_version,
     cleanup_version,
 )
+from pyobo.utils.ver import get_version, get_version_pins
 
 MOCK_PYOBO_VERSION_PINS = '{"ncbitaxon": "2024-07-03", "vo":"2024-04-09", "chebi":"235", "bfo":5}'
 FAULTY_MOCK_PYOBO_VERSION_PINS = "{'ncbitaxon': '2024-07-03'}"
